+++ src/pas/plugins/authomatic/integration/__init__.py
@@ -1,0 +1,2 @@
+from .restapi import RestAPIAdapter  # noQA
+from .zope import ZopeRequestAdapter  # noQA